Separated methods into a different file Update the adding of profile pictures to work properly
