+++ src/main/java/seedu/address/model/person/DisplayPic.java
@@ -3,31 +3,18 @@ package seedu.address.model.person;
 import static java.util.Objects.requireNonNull;
 import static seedu.address.commons.util.AppUtil.checkArgument;
 
-import java.awt.image.BufferedImage;
-import java.io.File;
-import java.io.IOException;
-import java.io.InputStream;
-import java.nio.file.Paths;
-
-import javax.imageio.ImageIO;
-
-import seedu.address.MainApp;
+import seedu.address.commons.core.Messages;
 import seedu.address.commons.exceptions.IllegalValueException;
 import seedu.address.commons.util.FileUtil;
+import seedu.address.storage.DisplayPicStorage;
 
 /**
  * Represents the filepath of a Person's displayPic in the address book.
- * Guarantees: immutable; is valid as declared in {@link #isValidPath(String)}
  */
 public class DisplayPic {
 
-    public static final String MESSAGE_DISPLAY_PIC_NONEXISTENT_CONSTRAINTS =
-            "The filepath should lead to a file that exists.";
-    public static final String MESSAGE_DISPLAY_PIC_NOT_IMAGE =
-            "The filepath should point to a valid image file.";
-    public static final String DEFAULT_DISPLAY_PIC = "/images/displayPic/default.png";
-    public static final String DEFAULT_IMAGE_LOCATION = "/images/displayPic/";
-    public static final String SAVE_LOCATION = "/src/main/resources/images/displayPic/";
+    public static final String DEFAULT_DISPLAY_PIC = "src/main/resources/images/displayPic/default.png";
+    public static final String DEFAULT_IMAGE_LOCATION = "src/main/resources/images/displayPic/";
 
     public final String value;
 
@@ -43,13 +30,13 @@ public class DisplayPic {
     public DisplayPic(String name, String filePath) throws IllegalValueException {
         requireNonNull(filePath);
         String trimmedFilePath = filePath.trim();
-        checkArgument(isValidPath(trimmedFilePath), MESSAGE_DISPLAY_PIC_NONEXISTENT_CONSTRAINTS);
-        checkArgument(isValidImage(trimmedFilePath), MESSAGE_DISPLAY_PIC_NOT_IMAGE);
+        checkArgument(DisplayPicStorage.isValidPath(trimmedFilePath),
+                Messages.MESSAGE_DISPLAY_PIC_NONEXISTENT_CONSTRAINTS);
+        checkArgument(DisplayPicStorage.isValidImage(trimmedFilePath), Messages.MESSAGE_DISPLAY_PIC_NOT_IMAGE);
 
 
         String fileType = FileUtil.getFileType(trimmedFilePath);
-        if (saveDisplayPic(name.trim(), trimmedFilePath, fileType)) {
-            //this.value = name.trim() + '.' + fileType;
+        if (DisplayPicStorage.saveDisplayPic(name.trim(), trimmedFilePath, fileType)) {
             this.value = DEFAULT_IMAGE_LOCATION + name.trim() + '.' + fileType;
         } else {
             this.value = DEFAULT_DISPLAY_PIC;
@@ -58,67 +45,9 @@ public class DisplayPic {
 
     public DisplayPic(String filePath) {
         requireNonNull(filePath);
-        checkArgument(isValidPath(filePath), MESSAGE_DISPLAY_PIC_NONEXISTENT_CONSTRAINTS);
-        checkArgument(isValidImage(filePath), MESSAGE_DISPLAY_PIC_NOT_IMAGE);
+        checkArgument(DisplayPicStorage.isValidPath(filePath), Messages.MESSAGE_DISPLAY_PIC_NONEXISTENT_CONSTRAINTS);
+        checkArgument(DisplayPicStorage.isValidImage(filePath), Messages.MESSAGE_DISPLAY_PIC_NOT_IMAGE);
         this.value = filePath;
-    }
-
-    /**
-     * Tries to save a copy of the image provided by the user into a default location.
-     * @param name the name of the new image file
-     * @param filePath the location of the current image file
-     * @param fileType the file extension of the current image file
-     * @return whether the image was successfully copied
-     */
-    public boolean saveDisplayPic(String name, String filePath, String fileType) {
-        try {
-            BufferedImage image = ImageIO.read(new File(filePath));
-            FileUtil.copyImage(image, fileType, Paths.get(".").toAbsolutePath().normalize().toString()
-                    + SAVE_LOCATION + name.trim() + '.' + fileType);
-            return true;
-        } catch (IOException | IllegalValueException exp) {
-            return false;
-        }
-    }
-
-    /**
-     * Returns true if a given string points to a valid file.
-     */
-    public static boolean isValidPath(String test) {
-        if (MainApp.class.getResourceAsStream(test) == null) {
-            File file = new File(test);
-            return FileUtil.isFileExists(file);
-        } else {
-            return true;
-        }
-    }
-
-    /**
-     * Checks if the image file provided can be opened properly as an image
-     * @param test is a filepath to an image file
-     * @return if the filePath it is pointing to is am image file that can be opened
-     */
-    public static boolean isValidImage(String test) {
-        try {
-            InputStream imageStream = ImageIO.class.getResourceAsStream(test);
-            if (imageStream == null) {
-                try {
-                    BufferedImage image = ImageIO.read(new File(test));
-                    return image != null;
-                } catch (IOException e3) {
-                    return false;
-                }
-            }
-            BufferedImage image = ImageIO.read(imageStream);
-            return image != null;
-        } catch (IOException e) {
-            try {
-                BufferedImage image = ImageIO.read(new File(test));
-                return image != null;
-            } catch (IOException e2) {
-                return false;
-            }
-        }
     }
 
     @Override
